@@ -2,44 +2,36 @@
 
 class Program
 {
-    // <mainandhandler>
-    static async Task<int> Main(string[] args)
+    // <asyncaction>
+    static Task<int> Main(string[] args)
     {
-        int returnCode = 0;
+        Option<string> urlOption = new("--url", "A URL.");
+        RootCommand rootCommand = new("Handle termination example") { urlOption };
 
-        var urlOption = new Option<string>("--url", "A URL.");
+        rootCommand.SetAction((ParseResult parseResult, CancellationToken cancellationToken) =>
+        {
+            string? urlOptionValue = parseResult.GetValue(urlOption);
+            return DoRootCommand(urlOptionValue, cancellationToken);
+        });
 
-        var rootCommand = new RootCommand("Handle termination example");
-        rootCommand.Add(urlOption);
-
-        rootCommand.SetHandler(async (context) =>
-            {
-                string? urlOptionValue = context.ParseResult.GetValueForOption(urlOption);
-                var token = context.GetCancellationToken();
-                returnCode = await DoRootCommand(urlOptionValue, token);
-            });
-
-        await rootCommand.InvokeAsync(args);
-
-        return returnCode;
+        return rootCommand.Parse(args).InvokeAsync();
     }
 
     public static async Task<int> DoRootCommand(
         string? urlOptionValue, CancellationToken cancellationToken)
     {
+        using HttpClient httpClient = new();
+
         try
         {
-            using (var httpClient = new HttpClient())
-            {
-                await httpClient.GetAsync(urlOptionValue, cancellationToken);
-            }
+            await httpClient.GetAsync(urlOptionValue, cancellationToken);
             return 0;
         }
         catch (OperationCanceledException)
         {
-            Console.Error.WriteLine("The operation was aborted");
+            await Console.Error.WriteLineAsync("The operation was aborted");
             return 1;
         }
     }
-    // </mainandhandler>
+    // </asyncaction>
 }
